requirements.txt
fitparse
sqlalchemy
psycopg2-binary
click
rich
numpy
python-fitparse 
python-dotenv